% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colored_Facet.R
\name{colored_Facet}
\alias{colored_Facet}
\title{colored_Facet}
\usage{
colored_Facet(
  dataframe = data.frame(y_values = c(1:1000), genotype = factor(rep(c("WT", "Mutant"),
    500), levels = c("WT", "Mutant")), timepoint = rep(c("early", "early", "late",
    "late"), 250)),
  ggplot_fun = function(df) {
     ggplot2::ggplot(df, ggplot2::aes(color = genotype, y
    = y_values, fill = timepoint, x = timepoint)) + ggplot2::geom_boxplot() +
    ggplot2::facet_wrap(~get(faceting_group), drop = T) + ggplot2::guides(color = FALSE)
    + ggplot2::theme(strip.background = ggplot2::element_rect(fill = strip_color))
 },
  faceting_group = "genotype",
  nrow = 1,
  strip_colors = c(WT = "blue", Mutant = "orange"),
  legend_proportion = 0.2
)
}
\arguments{
\item{dataframe}{The dataframe with data to plot}

\item{ggplot_fun}{A function to make a non-faceted plot from the data in the dataframe}

\item{faceting_group}{The grouping variable (ie column name) used for making the separate plots}

\item{nrow}{The number of rows to make for the plots}

\item{strip_colors}{A vector of colors equal to the number of variables in the faceting_group. Each color should have a name equal to a faceting variable. The order of colors will determine order of plots.}

\item{legend_proportion}{Proportion of the horizontal plot area taken by the shared legend}
}
\value{
a ggplot in which facet group strips are colored
}
\description{
Given a dataframe, a faceting group, and a function to make an overall plot, this returns a faceted plot with colored banners over each subplot.
}
